T3_Recommandation/src/neural_network/requirements.txt
torch
pandas
numpy
scikit-learn
psycopg2-binary
python-dotenv
tqdm
pyyaml
